Renamed editor-modal; js_only.
