+++ assets/prototype/domain/eventEditor/context/TicketProvider/useTicketEditorId.ts
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { EntityId } from '../../data/types';
-import { EditorId } from '../../../../application/ui/components/layout/editor-modal/useEditorModalState/types';
+import { EditorId } from '../../../../application/ui/components/layout/editorModal/useEditorModalState/types';
 
 /**
  * @function
